touchscreen: gt9896: Make the driver init early
Currently we are using Goodix gt9896 touchscreen hardware which means it has
to use Goodix gt9896 drivers as well. But when it comes to initialization
the FTS K6 touchscreen driver does it earlier than the Goodix gt9896 driver.
This causes an early error when making a request to the vdd GPIO and the
touchscreen cannot be used. To fix this simply increase the initialization
level for the Goodix gt9896 driver slightly in the goodix_i2c_init()
function using device_initcall_sync() for earlier initialization of
the FTS K6 touchscreen driver.

Signed-off-by: Fiqri Ardyansyah <fiqri0927936@gmail.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>